organize cohort_handling
